Show loading for each widgets on Dashboard independently
